Rename db folder
